Makefile and build system: simplifications
The former Makefile was used to create on the fly necessary .cpp files from .cu ones. Now the .cpp files are part of the distribution to avoid confusions.

The Makefile now writes the intermediate build products in matlab/mex/.build. It also does not remove them after the compilation is done, in order to save time when the package is compiled again.