Speed-up loading of the DagNN models.
Rebuilding the network after addition of each single layer
slows down loading of large networks signifficantly. Observable
mainly on slower machines. Solved by adding an option to skip
the rebuild step and exception handling in the loadobj function
such that the behaviour does not change to previous version even
when an exception is thrown.