Small improvements of the print function.
Additionally added the posibility to specify arguments passed to dot.
Possibility to change the format even for temporary files.
Stores the .dot file to the same path as figurepath.
Cleans up the temporary files.
Added a super simple function which just calls print on a MNIST net.
Dot output printed only if not empty.